Activate PGI compiler bug workaround for Emu
Most GraphBLAS functions are type-generic, they can run on int, loat,
long, char, etc.

In some cases, the build system generates multiple versions of each
function from a template, and compiles each one separately (yielding
foo_INT64, foo_FLOAT, etc.). This is part of the reason why the build
takes so long!

In other cases, a single function handles all of the types. Type
information is passed down the call tree along with each data structure.
When the function needs to create a temporary object of the type, it
uses a variable-length array like so:

GB_void tmp[zsize];

Where zsize is a variable specifying the size of our mystery type.

On most systems, you can simply grow the stack to make room for the VLA.

But on Emu we have cactus stacks. We call malloc() to create
storage for the VLA, with a corresponding free() when the VLA goes out
of scope. This is very slow.

Interestingly, GraphBLAS uses a macro at the location of each VLA, which
was intended to work around a bug in the PGI compiler related to VLA's.
It replaces the VLA with a fixed-size array. This would be advantageous
for us; since the array size is a compile-time constant, we can put it
on the stack and eliminate the malloc/free calls.

So I'm going to activate the PGI bug macro and see how well it works.
Only downside is that it wastes memory.